Time: 124 ms (99.52%), Space: 69.7 MB (83.22%) - LeetHub

+++ 304-range-sum-query-2d-immutable/304-range-sum-query-2d-immutable.java
@@ -1,8 +1,8 @@
 class NumMatrix {
-    long[][] arr;
+    int[][] arr;
 
     public NumMatrix(int[][] matrix) {
-        arr = new long[matrix.length+1][matrix[0].length+1];
+        arr = new int[matrix.length+1][matrix[0].length+1];
         
         for (int i = 1; i < matrix.length+1; i++) {
             for (int j = 1; j < matrix[0].length+1; j++) {
@@ -12,7 +12,7 @@ class NumMatrix {
     }
     
     public int sumRegion(int row1, int col1, int row2, int col2) {
-        return (int)(arr[row2+1][col2+1] - arr[row2+1][col1] - arr[row1][col2+1] + arr[row1][col1]);
+        return arr[row2+1][col2+1] - arr[row2+1][col1] - arr[row1][col2+1] + arr[row1][col1];
     }
 }
